+++ packages/cli/src/client-config/client_config_lifecycle_handler.ts
@@ -1,0 +1,34 @@
+import {
+  ClientConfigFormat,
+  getClientConfigPath,
+} from '@aws-amplify/client-config';
+import { ClientConfigGeneratorAdapter } from './client_config_generator_adapter.js';
+import { BackendIdentifier } from '@aws-amplify/deployed-backend-client';
+import fsp from 'fs/promises';
+
+/**
+ * Life cycle handler for client config files written to file system
+ */
+export class ClientConfigLifecycleHandler {
+  /**
+   * Instantiate a new life cycle handler which proxy requests to ClientConfigGeneratorAdapter
+   */
+  constructor(
+    private clientConfigGeneratorAdapter: ClientConfigGeneratorAdapter,
+    private readonly outDir?: string,
+    private readonly format?: ClientConfigFormat
+  ) {}
+
+  generateClientConfigFile = async (backendIdentifier: BackendIdentifier) => {
+    await this.clientConfigGeneratorAdapter.generateClientConfigToFile(
+      backendIdentifier,
+      this.outDir,
+      this.format
+    );
+  };
+
+  deleteClientConfigFile = async () => {
+    const path = await getClientConfigPath(this.outDir, this.format);
+    await fsp.rm(path);
+  };
+}