+++ packages/cli/src/commands/sandbox/sandbox_event_handler_factory.ts
@@ -1,5 +1,4 @@
 import { SandboxBackendIdentifier } from '@aws-amplify/platform-core';
-import { ClientConfigGeneratorAdapter } from '../../client-config/client_config_generator_adapter.js';
 import { SandboxEventHandlerCreator } from './sandbox_command.js';
 
 /**
@@ -10,25 +9,26 @@ export class SandboxEventHandlerFactory {
    * Creates a SandboxEventHandlerFactory
    */
   constructor(
-    private clientConfigGeneratorAdapter: ClientConfigGeneratorAdapter,
     private getBackendIdentifier: (
       appName?: string
     ) => Promise<SandboxBackendIdentifier>
   ) {}
   getSandboxEventHandlers: SandboxEventHandlerCreator = ({
     appName,
-    clientConfigOutDir,
-    format,
+    clientConfigLifecycleHandler,
   }) => {
     return {
       successfulDeployment: [
         async () => {
           const backendIdentifier = await this.getBackendIdentifier(appName);
-          await this.clientConfigGeneratorAdapter.generateClientConfigToFile(
-            backendIdentifier,
-            clientConfigOutDir,
-            format
+          await clientConfigLifecycleHandler.generateClientConfigFile(
+            backendIdentifier
           );
+        },
+      ],
+      successfulDeletion: [
+        async () => {
+          await clientConfigLifecycleHandler.deleteClientConfigFile();
         },
       ],
     };